requirements.txt
python-dotenv==1.0.1
watchdog==4.0.0
Flask==3.0.3
Flask-WTF==1.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6 
itsdangerous==2.2.0
Flask-Mail2==1.0.2
